Split container gui in multiple files
