+++ fmu_manipulation_toolbox/gui/fmueditor/__main__.py
@@ -585,5 +585,3 @@ def main():
 
 if __name__ == "__main__":
     main()
-
-
